[#2] 공통 컴포넌트 제작
- AddressBoxCard 분리 제작

#2
