+++ lib/presentation/screens/test/box_card_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../../core/constants/color.dart';
+import '../../widgets/address_box_card.dart';
 import '../../widgets/custom_box_card.dart';
 
 class BoxCardTest extends StatefulWidget {
